Beta Test/Report 3
My mom played this game and liked it overall. I noticed that she tried
to eat the sandwich instead of giving it to the shark so I made the
sandwich a fish sandwich so you couldn’t eat it. Other than a that and
a few spelling errors, there were not mistakes.

Problems: player eating sandwich, spelling.

Solution: turned the ham sandwich into a fish sandwich. Corrected the
spelling errors.